+++ src/main/java/backend/mcsvventas/models/dtos/request/VentaRequestDto.java
@@ -4,6 +4,6 @@ import backend.mcsvventas.models.documents.DetalleVenta;
 
 import java.util.List;
 
-public record VentaRequestDto(Long clientId,
+public record VentaRequestDto(Integer clientId,
                               List<DetalleVentaRequestDto> details) {
 }
